+++ changelog.md
@@ -5,3 +5,4 @@
 - Added a button to export mods
 - Added a custom layer to view the button
 - Added a button to show the layer (right to the Geode button)
+- Added a button to import mods (not implemented yet)